+++ pwsh/dev/devAzGovVizParallel.ps1
@@ -365,7 +365,7 @@ Param
     $Product = 'AzGovViz',
 
     [string]
-    $AzAPICallVersion = '1.1.72',
+    $AzAPICallVersion = '1.1.77',
 
     [string]
     $ProductVersion = '6.3.1',
@@ -781,12 +781,12 @@ Write-Host " Initialize 'AzAPICall' succeeded" -ForegroundColor Green
 #EndRegion initAZAPICall
 
 #region required AzAPICall version
-if (-not ([System.Version]"$($azapicallConf['htParameters'].azAPICallModuleVersion)" -ge [System.Version]'1.1.72')) {
+if (-not ([System.Version]"$($azapicallConf['htParameters'].azAPICallModuleVersion)" -ge [System.Version]'1.1.77')) {
     Write-Host 'AzAPICall version check failed -> https://aka.ms/AzAPICall; https://www.powershellgallery.com/packages/AzAPICall'
-    throw 'This version of Azure Governance Visualizer requires AzAPICall module version 1.1.72 or greater'
+    throw 'This version of Azure Governance Visualizer requires AzAPICall module version 1.1.77 or greater'
 }
 else {
-    Write-Host "AzAPICall module version requirement check succeeded: 1.1.72 or greater - current: $($azapicallConf['htParameters'].azAPICallModuleVersion) " -ForegroundColor Green
+    Write-Host "AzAPICall module version requirement check succeeded: 1.1.77 or greater - current: $($azapicallConf['htParameters'].azAPICallModuleVersion) " -ForegroundColor Green
 }
 #endregion required AzAPICall version