+++ history.md
@@ -41,7 +41,7 @@ __Changes__ (2023-Mar-30 / 6.1.0 Major)
   * the new file for version check is version.json
 * Add updatedBy/updatedOn metadata for RBAC Role assignments
 * Add least privilege check for script execution in the context of a user for Azure Resource permissions (at this time it only checks permissions on the target Management Group Id) - best practice is to execute as a Service Principal with least privilege
-* Use [AzAPICall](https://aka.ms/AzAPICall) PowerShell module version 1.1.72
+* Use [AzAPICall](https://aka.ms/AzAPICall) PowerShell module version 1.1.77
   * add AzAPICall version information in AzAPICall outputs
   * if context is user then get the users objectId (required for least privilege check)